+++ src/components/Button/Button.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { select, boolean, text } from '@storybook/addon-knobs'
 import { ReactComponent as SunIcon } from '../../assets/svg/Sun.svg'
 import Button from './Button'
-import styles from '../../styles/storybook/index.module.scss'
+import styles from '../../storybook/index.module.scss'
 import { centerDecorator } from '../../utils/storybook/decorators'
 
 export default {