+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Unreleased
+### Breaking Changes :boom:
+- Added label prop to textField
+
+### Changed
+- Changed textfield to be uncontrolled
 
 ## 0.1.3 - 01-09-2020
 ### Added