+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 - Added onCellClick prop to table rows
+
 ## 0.1.3 - 01-09-2020
 ### Added
 - New Table.