+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 - New @anyvision/anv-icon section.
+- Added triggerRowClick to table cell
 
 ## 0.1.3 - 01-09-2020
 ### Added